@@ -1,4 +1,4 @@
-package com.robotemi.sdk.sample
+package com.seotepa.infobotApp
 
 import android.Manifest
 import android.annotation.SuppressLint
@@ -9,10 +9,8 @@ import android.content.DialogInterface
 import android.content.Intent
 import android.content.IntentFilter
 import android.content.pm.PackageManager
-import android.content.res.AssetFileDescriptor
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
-import android.media.MediaPlayer
 import android.os.Bundle
 import android.os.Environment
 import android.os.RemoteException
@@ -28,11 +26,16 @@ import android.view.inputmethod.InputMethodManager
 import android.widget.AdapterView
 import android.widget.AdapterView.OnItemClickListener
 import android.widget.ArrayAdapter
-import android.widget.SeekBar
-import android.widget.SeekBar.OnSeekBarChangeListener
 import android.widget.Toast
+import androidx.activity.compose.setContent
 import androidx.annotation.CheckResult
 import androidx.appcompat.app.AppCompatActivity
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.core.app.ActivityCompat
 import com.google.gson.Gson
 import com.robotemi.sdk.*
@@ -63,25 +66,25 @@ import com.robotemi.sdk.navigation.model.SafetyLevel
 import com.robotemi.sdk.navigation.model.SpeedLevel
 import com.robotemi.sdk.permission.OnRequestPermissionResultListener
 import com.robotemi.sdk.permission.Permission
+import com.seotepa.infobotApp.databinding.ActivityFaceBinding
+import com.seotepa.infobotApp.databinding.ActivityMainBinding
+import com.seotepa.infobotApp.databinding.GroupAppAndPermissionBinding
+import com.seotepa.infobotApp.databinding.GroupButtonsBinding
+import com.seotepa.infobotApp.databinding.GroupMapAndMovementBinding
+import com.seotepa.infobotApp.databinding.GroupSettingsAndStatusBinding
+import com.seotepa.infobotApp.navigation.AppNavigation
 import com.robotemi.sdk.sequence.OnSequencePlayStatusChangedListener
 import com.robotemi.sdk.sequence.SequenceModel
 import com.robotemi.sdk.telepresence.CallState
-import com.robotemi.sdk.telepresence.LinkBasedMeeting
 import com.robotemi.sdk.telepresence.Participant
 import com.robotemi.sdk.voice.ITtsService
 import com.robotemi.sdk.voice.model.TtsVoice
-import kotlinx.android.synthetic.main.activity_main.*
-import kotlinx.android.synthetic.main.group_app_and_permission.*
-import kotlinx.android.synthetic.main.group_buttons.*
-import kotlinx.android.synthetic.main.group_map_and_movement.*
-import kotlinx.android.synthetic.main.group_resources.*
-import kotlinx.android.synthetic.main.group_settings_and_status.*
 import java.io.File
 import java.io.FileOutputStream
 import java.io.IOException
 import java.util.*
 import java.util.concurrent.Executors
-import kotlin.concurrent.thread
+import com.seotepa.infobotApp.ui.theme.SdkTheme
 
 
 class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
@@ -110,6 +113,7 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
 
     private val assistantReceiver = AssistantChangeReceiver()
 
+
     private val telepresenceStatusChangedListener: OnTelepresenceStatusChangedListener by lazy {
         object : OnTelepresenceStatusChangedListener("") {
             override fun onTelepresenceStatusChanged(callState: CallState) {
@@ -119,14 +123,34 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
     }
 
     @SuppressLint("SetTextI18n")
+    private lateinit var bindingMain: ActivityMainBinding
+    private lateinit var bindingFace: ActivityFaceBinding
+    private lateinit var bindingGroupButtons: GroupButtonsBinding
+    private lateinit var bindingGroupSettingsStatus: GroupSettingsAndStatusBinding
+    private lateinit var bindingMapMovement: GroupMapAndMovementBinding
+    private lateinit var bindingAppPermissions: GroupAppAndPermissionBinding
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContent {
+            SdkTheme {
+                // A surface container using the 'background' color from the theme
+                Surface(modifier = Modifier.fillMaxSize(), color = MaterialTheme.colorScheme.background) {
+                    AppNavigation()
+                }
+            }
+        }
+        /*setContentView(R.layout.activity_main)*/
         window.setSoftInputMode(WindowManager.LayoutParams.SOFT_INPUT_STATE_HIDDEN)
         verifyStoragePermissions(this)
         robot = getInstance()
         initOnClickListener()
-        tvLog.movementMethod = ScrollingMovementMethod.getInstance()
+        bindingFace = ActivityFaceBinding.inflate(layoutInflater)
+        bindingGroupButtons = GroupButtonsBinding.inflate(layoutInflater)
+        bindingGroupSettingsStatus = GroupSettingsAndStatusBinding.inflate(layoutInflater)
+        bindingMapMovement = GroupMapAndMovementBinding.inflate(layoutInflater)
+        bindingAppPermissions = GroupAppAndPermissionBinding.inflate(layoutInflater)
+        bindingFace.tvLog.movementMethod = ScrollingMovementMethod.getInstance()
         robot.addOnRequestPermissionResultListener(this)
         robot.addOnTelepresenceEventChangedListener(this)
         robot.addOnFaceRecognizedListener(this)
@@ -145,11 +169,6 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
             printLog("Override tts")
             tts = TextToSpeech(this, this)
             robot.setTtsService(this)
-        }
-        val greetModeState = intent.extras?.getInt(SdkConstants.INTENT_ACTION_GREET_MODE_STATE)
-        if (greetModeState != null) {
-            tvGreetMode.text =
-                "Greet Mode -> ${OnGreetModeStateChangedListener.State.fromValue(greetModeState)}"
         }
         debugReceiver = TemiBroadcastReceiver()
         registerReceiver(debugReceiver, IntentFilter(TemiBroadcastReceiver.ACTION_DEBUG))
@@ -264,278 +283,7 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
     }
 
     private fun initOnClickListener() {
-        btnGroupSystem.setOnCheckedChangeListener { _, isChecked ->
-            if (isChecked) {
-                group_settings_and_status.visibility = View.VISIBLE
-                btnGroupSystem.isEnabled = false
-                btnGroupNavigation.isChecked = false
-                btnGroupPermission.isChecked = false
-                btnGroupResources.isChecked = false
-            } else {
-                group_settings_and_status.visibility = View.GONE
-                btnGroupSystem.isEnabled = true
-            }
-        }
-        btnGroupNavigation.setOnCheckedChangeListener { _, isChecked ->
-            if (isChecked) {
-                group_map_and_movement.visibility = View.VISIBLE
-                btnGroupSystem.isChecked = false
-                btnGroupNavigation.isEnabled = false
-                btnGroupPermission.isChecked = false
-                btnGroupResources.isChecked = false
-            } else {
-                group_map_and_movement.visibility = View.GONE
-                btnGroupNavigation.isEnabled = true
-            }
-        }
-        btnGroupPermission.setOnCheckedChangeListener { _, isChecked ->
-            if (isChecked) {
-                group_app_and_permission.visibility = View.VISIBLE
-                btnGroupSystem.isChecked = false
-                btnGroupNavigation.isChecked = false
-                btnGroupPermission.isEnabled = false
-                btnGroupResources.isChecked = false
-            } else {
-                group_app_and_permission.visibility = View.GONE
-                btnGroupPermission.isEnabled = true
-            }
-        }
-        btnGroupResources.setOnCheckedChangeListener { _, isChecked ->
-            if (isChecked) {
-                group_resources.visibility = View.VISIBLE
-                btnGroupSystem.isChecked = false
-                btnGroupNavigation.isChecked = false
-                btnGroupPermission.isChecked = false
-                btnGroupResources.isEnabled = false
-            } else {
-                group_resources.visibility = View.GONE
-                btnGroupResources.isEnabled = true
-            }
-        }
 
-        val mediaPlayer = MediaPlayer()
-
-        btnGroupSystem.isChecked = true
-
-        btnSpeak.setOnClickListener { speak() }
-        btnSaveLocation.setOnClickListener { saveLocation() }
-        btnGoTo.setOnClickListener { goTo() }
-
-        btnStopMovement.setOnClickListener { stopMovement() }
-        btnFollow.setOnClickListener { followMe() }
-        btnskidJoy.setOnClickListener { skidJoy() }
-        btnTiltAngle.setOnClickListener { tiltAngle() }
-        btnTiltBy.setOnClickListener { tiltBy() }
-        btnTurnBy.setOnClickListener { turnBy() }
-        btnBatteryInfo.setOnClickListener { getBatteryData() }
-        btnSavedLocations.setOnClickListener { savedLocationsDialog() }
-        btnCallOwner.setOnClickListener { callOwner() }
-        btnStopCall.setOnClickListener {
-            if (!requestPermissionIfNeeded(Permission.MEETINGS, REQUEST_CODE_NORMAL)) {
-                stopCall()
-            }
-        }
-        btnPublish.setOnClickListener { publishToActivityStream() }
-        btnHideTopBar.setOnClickListener { hideTopBar() }
-        btnShowTopBar.setOnClickListener { showTopBar() }
-        btnDisableWakeup.setOnClickListener { disableWakeup() }
-        btnEnableWakeup.setOnClickListener { enableWakeup() }
-        btnToggleNavBillboard.setOnClickListener { toggleNavBillboard() }
-        btnTogglePrivacyModeOn.setOnClickListener { privacyModeOn() }
-        btnTogglePrivacyModeOff.setOnClickListener { privacyModeOff() }
-        btnGetPrivacyMode.setOnClickListener { getPrivacyModeState() }
-        btnEnableHardButtons.setOnClickListener { enableHardButtons() }
-        btnDisableHardButtons.setOnClickListener { disableHardButtons() }
-        btnIsHardButtonsDisabled.setOnClickListener { isHardButtonsEnabled() }
-        btnGetOSVersion.setOnClickListener { getOSVersion() }
-        btnCheckFace.setOnClickListener { requestFace() }
-        btnCheckMap.setOnClickListener { requestMap() }
-        btnCheckSettings.setOnClickListener { requestSettings() }
-        btnCheckSequence.setOnClickListener { requestSequence() }
-        btnCheckMeetings.setOnClickListener { requestMeetings() }
-        btnCheckAllPermission.setOnClickListener { requestAll() }
-        btnStartFaceRecognition.setOnClickListener { startFaceRecognition() }
-        btnStopFaceRecognition.setOnClickListener { stopFaceRecognition() }
-        btnTestFaceRecognition.setOnClickListener { testFaceRecognition() }
-        btnSetUserInteractionON.setOnClickListener {
-            val ret = robot.setInteractionState(true)
-            Log.d("MainActivity", "Set user interaction $ret")
-            mediaPlayer.setVolume(1f, 1f)
-            mediaPlayer.isLooping = false
-            mediaPlayer.setOnCompletionListener {
-                robot.setInteractionState(false)
-            }
-            if (!mediaPlayer.isPlaying) {
-                val descriptor: AssetFileDescriptor = assets.openFd("Lorem-ipsum.mp3")
-                mediaPlayer.setDataSource(
-                    descriptor.fileDescriptor,
-                    descriptor.startOffset,
-                    descriptor.length
-                )
-                descriptor.close()
-                mediaPlayer.prepare()
-                mediaPlayer.start()
-            }
-        }
-        btnSetUserInteractionOFF.setOnClickListener {
-            robot.setInteractionState(false)
-            mediaPlayer.stop()
-            mediaPlayer.reset()
-        }
-        btnSetGoToSpeed.setOnClickListener { setGoToSpeed() }
-        btnSetGoToSafety.setOnClickListener { setGoToSafety() }
-        btnToggleTopBadge.setOnClickListener { toggleTopBadge() }
-        btnToggleDetectionMode.setOnClickListener { toggleDetectionMode() }
-        btnToggleAutoReturn.setOnClickListener { toggleAutoReturn() }
-        btnTrackUser.setOnClickListener { toggleTrackUser() }
-        btnGetVolume.setOnClickListener { getVolume() }
-        btnSetVolume.setOnClickListener { setVolume() }
-        btnRequestToBeKioskApp.setOnClickListener { requestToBeKioskApp() }
-        btnStartDetectionModeWithDistance.setOnClickListener { startDetectionWithDistance() }
-        btnFetchSequence.setOnClickListener { getAllSequences() }
-        btnPlayFirstSequence.setOnClickListener { playFirstSequence() }
-        btnPlayFirstSequenceWithoutPlayer.setOnClickListener { playFirstSequenceWithoutPlayer() }
-        btnFetchMap.setOnClickListener { getMap() }
-        btnClearLog.setOnClickListener { clearLog() }
-        btnNlu.setOnClickListener { startNlu() }
-        btnGetAllContacts.setOnClickListener { getAllContacts() }
-        btnGoToPosition.setOnClickListener { goToPosition() }
-        btnStartTelepresenceToCenter.setOnClickListener { startTelepresenceToCenter() }
-        btnStartMeeting.setOnClickListener { startMeeting() }
-        btnCreateLinkBasedMeeting.setOnClickListener {
-            if (requestPermissionIfNeeded(Permission.MEETINGS, REQUEST_CODE_NORMAL)) {
-                // Permission not granted yet.
-            } else {
-                val request = LinkBasedMeeting(
-                    topic = "temi Demo Meeting",
-                    availability = LinkBasedMeeting.Availability(
-                        start = Date(),
-                        end = Date(Date().time + 86400000),
-                        always = false,
-                    ),
-                    limit = LinkBasedMeeting.Limit(
-                        callDuration = LinkBasedMeeting.CallDuration.MINUTE_10,
-                        usageLimit = LinkBasedMeeting.UsageLimit.NO_LIMIT,
-                    ),
-                    permission = LinkBasedMeeting.Permission.DEFAULT,
-                    security = LinkBasedMeeting.Security(
-                        password = "1122334455", // Should use a 1 to 10-digits password.
-                        hasPassword = false
-                    )
-                )
-                thread {
-                    val (code, linkUrl) = robot.createLinkBasedMeeting(request)
-                    printLog("Link create request, response code $code, link $linkUrl")
-                }
-            }
-        }
-        btnStartPage.setOnClickListener { startPage() }
-        btnRestart.setOnClickListener { restartTemi() }
-        btnGetMembersStatus.setOnClickListener { getMembersStatus() }
-        btnRepose.setOnClickListener { repose() }
-        btnGetMapList.setOnClickListener { getMapListBtn() }
-        btnLoadMap.setOnClickListener { loadMap() }
-        btnLoadMapToCache.setOnClickListener { loadMapToCache() }
-        btnLoadMapOffline.setOnClickListener { loadMap(false, null, true) }
-        btnLoadMapWithoutUI.setOnClickListener {
-            loadMap(
-                false,
-                null,
-                offline = false,
-                withoutUI = true
-            )
-        }
-        btnLock.setOnClickListener { lock() }
-        btnUnlock.setOnClickListener { unlock() }
-        btnMuteAlexa.setOnClickListener { muteAlexa() }
-        btnShutdown.setOnClickListener { shutdown() }
-        btnLoadMapWithPosition.setOnClickListener { loadMapWithPosition() }
-        btnLoadMapWithReposePosition.setOnClickListener { loadMapWithReposePosition() }
-        btnLoadMapWithRepose.setOnClickListener { loadMapWithRepose() }
-        btnSetSoundMode.setOnClickListener { setSoundMode() }
-        btnSetHardBtnMainMode.setOnClickListener { setHardBtnMainMode() }
-        btnToggleHardBtnPower.setOnClickListener { toggleHardBtnPower() }
-        btnToggleHardBtnVolume.setOnClickListener { toggleHardBtnVolume() }
-        btnGetNickName.setOnClickListener { getNickName() }
-        btnSetMode.setOnClickListener { setMode() }
-        btnGetMode.setOnClickListener { getMode() }
-        btnToggleKioskMode.setOnClickListener { toggleKiosk() }
-        btnIsKioskModeOn.setOnClickListener { isKioskModeOn() }
-        btnEnabledLatinKeyboards.setOnClickListener { enabledLatinKeyboards() }
-        btnGetSupportedKeyboard.setOnClickListener { getSupportedLatinKeyboards() }
-        btnToggleGroundDepthCliff.setOnClickListener { toggleGroundDepthCliff() }
-        btnIsGroundDepthCliff.setOnClickListener { isGroundDepthCliffEnabled() }
-        btnHasCliffSensor.setOnClickListener { hasCliffSensor() }
-        btnSetCliffSensorMode.setOnClickListener { setCliffSensorMode() }
-        btnGetCliffSensorMode.setOnClickListener { getCliffSensorMode() }
-        btnSetHeadDepthSensitivity.setOnClickListener { setHeadDepthSensitivity() }
-        btnGetHeadDepthSensitivity.setOnClickListener { getHeadDepthSensitivity() }
-        btnToggleFrontTOF.setOnClickListener { toggleFrontTOF() }
-        btnIsFrontTOFEnabled.setOnClickListener { isFrontTOFEnabled() }
-        btnToggleBackTOF.setOnClickListener { toggleBackTOF() }
-        btnIsBackTOFEnabled.setOnClickListener { isBackTOFEnabled() }
-        btnMinimumObstacleDistance.setOnClickListener {
-            if (requestPermissionIfNeeded(Permission.SETTINGS, REQUEST_CODE_NORMAL)) {
-                return@setOnClickListener
-            }
-            if (robot.minimumObstacleDistance == -1) {
-                Toast.makeText(this, "Minimum Obstacle Distance settings is not supported on your robot.", Toast.LENGTH_SHORT).show()
-                return@setOnClickListener
-            }
-
-            if (groupMinimumObstacleDistance.visibility == View.GONE) {
-                 groupMinimumObstacleDistance.visibility = View.VISIBLE
-            }
-            val distance = robot.minimumObstacleDistance
-            textMinimumObstacleDistance.text = "$distance"
-            seekbarMinimumObstacleDistance.progress = distance.coerceIn(0, 100)
-            seekbarMinimumObstacleDistance.setOnSeekBarChangeListener(object : OnSeekBarChangeListener {
-                override fun onProgressChanged(
-                    seekBar: SeekBar?,
-                    progress: Int,
-                    fromUser: Boolean
-                ) {
-                    seekBar ?: return
-                    // Round to 5x
-                    val value = seekBar.progress / 5 * 5
-                    seekBar.progress = value
-                    textMinimumObstacleDistance.text = "$value"
-                }
-
-                override fun onStartTrackingTouch(seekBar: SeekBar?) {
-                    // nothing
-                }
-
-                override fun onStopTrackingTouch(seekBar: SeekBar?) {
-                    seekBar ?: return
-                    // Round to 5x
-                    val value = seekBar.progress / 5 * 5
-                    robot.minimumObstacleDistance = value
-                }
-
-            })
-        }
-        btnGetAllFloors.setOnClickListener { getAllFloors() }
-        btnLoadFloorAtElevator.setOnClickListener { loadFloorAtElevator() }
-        btnGetCurrentFloor.setOnClickListener {
-            getCurrentFloor()
-        }
-        btnGetTts.setOnClickListener { getTts() }
-        btnSetTts.setOnClickListener { setTts() }
-        btnSerial.setOnClickListener { startActivity(Intent(this, SerialActivity::class.java)) }
-        btnWebpage.setOnClickListener {
-            val intent =
-                Intent().setClassName("com.robotemi.browser", "com.robotemi.browser.MainActivity")
-            intent.putExtra("url", "https://github.com")
-            intent.putExtra("source", "intent")
-            intent.putExtra("navBar", "SHOW")
-            intent.putExtra("reset", "OFF")
-            try {
-                startActivity(intent)
-            } catch (e: Exception) {
-                printLog("Cannot launch browser, probably temi browser app not installed.")
-            }
-        }
     }
 
     private fun getCurrentFloor() {
@@ -829,7 +577,7 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
      * Have the robot speak while displaying what is being said.
      */
     private fun speak() {
-        val text = etSpeak.text.toString()
+        val text = bindingMain.etSpeak.text.toString()
         val languages = ArrayList<TtsRequest.Language>()
         TtsRequest.Language.values().forEach {
             language ->  languages.add(language)
@@ -856,7 +604,7 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
      */
     private fun saveLocation() {
         val location =
-            etSaveLocation.text.toString().lowercase().trim { it <= ' ' }
+            bindingMain.etSaveLocation.text.toString().lowercase().trim { it <= ' ' }
         val result = robot.saveLocation(location)
         if (result) {
             robot.speak(create("I've successfully saved the $location location.", true))
@@ -871,11 +619,11 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
      */
     private fun goTo() {
         for (location in robot.locations) {
-            if (location == etGoTo.text.toString().lowercase()
+            if (location == bindingMain.etGoTo.text.toString().lowercase()
                     .trim { it <= ' ' }
             ) {
                 robot.goTo(
-                    etGoTo.text.toString().lowercase().trim { it <= ' ' },
+                    bindingMain.etGoTo.text.toString().lowercase().trim { it <= ' ' },
                     backwards = false,
                     noBypass = false,
                     speedLevel = SpeedLevel.HIGH
@@ -911,12 +659,12 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
         val t = System.currentTimeMillis()
         val end = t + 500
         val speedX = try {
-            etX.text.toString().toFloat()
+            bindingMain.etX.text.toString().toFloat()
         } catch (e: Exception) {
             1f
         }
         val speedY = try {
-            etY.text.toString().toFloat()
+            bindingMain.etY.text.toString().toFloat()
         } catch (e: Exception) {
             0f
         }
@@ -932,7 +680,7 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
      */
     private fun tiltAngle() {
         val speed = try {
-            etDistance.text.toString().toFloat()
+            bindingMain.etDistance.text.toString().toFloat()
         } catch (e: Exception) {
             1f
         }
@@ -945,7 +693,7 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
      */
     private fun turnBy() {
         val speed = try {
-            etDistance.text.toString().toFloat()
+            bindingMain.etDistance.text.toString().toFloat()
         } catch (e: Exception) {
             1f
         }
@@ -957,7 +705,7 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
      */
     private fun tiltBy() {
         val speed = try {
-            etDistance.text.toString().toFloat()
+            bindingMain.etDistance.text.toString().toFloat()
         } catch (e: Exception) {
             1f
         }
@@ -1176,7 +924,7 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
     }
 
     override fun onDetectionStateChanged(state: Int) {
-        tvDetectionState.text =
+        bindingMain.tvDetectionState.text =
             "Detect State -> ${OnDetectionStateChangedListener.DetectionStatus.fromValue(state)}"
     }
 
@@ -1516,7 +1264,7 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
         ) {
             return
         }
-        var distanceStr = etDistance.text.toString()
+        var distanceStr = bindingMain.etDistance.text.toString()
         if (distanceStr.isEmpty()) distanceStr = "0"
         try {
             val distance = distanceStr.toFloat()
@@ -1533,12 +1281,12 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
             text +=
                 " -> $location :: ${distances[location]}\n"
         }
-        tvDistance.text = text
+        bindingMain.tvDistance.text = text
     }
 
     @SuppressLint("SetTextI18n")
     override fun onCurrentPositionChanged(position: Position) {
-        tvPosition.text =
+        bindingMain.tvPosition.text =
             "Position -> {${position.x}, ${position.y}, ${position.yaw}}, tilt: ${position.tiltAngle}"
     }
 
@@ -1570,7 +1318,7 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
     }
 
     override fun onDetectionDataChanged(detectionData: DetectionData) {
-        tvDetection.text = if (detectionData.isDetected) {
+        bindingMain.tvDetection.text = if (detectionData.isDetected) {
             "Detect -> angle ${detectionData.angle}, dist ${detectionData.distance}"
         } else {
             "No detection"
@@ -1589,7 +1337,7 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
             return
         }
         Thread {
-            allSequences = robot.getAllSequences(etNlu.text?.split(",") ?: emptyList())
+            allSequences = robot.getAllSequences(bindingMain.etNlu.text?.split(",") ?: emptyList())
             printLog("allSequences: ${allSequences.size}", false)
             val imageKeys: MutableList<String> = ArrayList()
             for ((_, _, _, imageKey) in allSequences) {
@@ -1643,7 +1391,7 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
     override fun onFaceRecognized(contactModelList: List<ContactModel>) {
         if (contactModelList.isEmpty()) {
             printLog("onFaceRecognized: User left")
-            imageViewFace.visibility = View.INVISIBLE
+            bindingMain.imageViewFace.visibility = View.INVISIBLE
             return
         }
 
@@ -1651,8 +1399,8 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
         if (!imageKey.isNullOrBlank()) {
             showFaceRecognitionImage(imageKey)
         } else {
-            imageViewFace.setImageResource(R.drawable.app_icon)
-            imageViewFace.visibility = View.VISIBLE
+            bindingMain.imageViewFace.setImageResource(R.drawable.app_icon)
+            bindingMain.imageViewFace.visibility = View.VISIBLE
         }
 
         for (contactModel in contactModelList) {
@@ -1685,8 +1433,8 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
         var text = ""
         if (contactModelList.isEmpty()) {
             text = "onContinuousFaceRecognized: User left"
-            imageViewFace.visibility = View.INVISIBLE
-            tvContinuousFace.text = text
+            bindingMain.imageViewFace.visibility = View.INVISIBLE
+            bindingMain.tvContinuousFace.text = text
             return
         }
 
@@ -1694,8 +1442,8 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
         if (!imageKey.isNullOrBlank()) {
             showFaceRecognitionImage(imageKey)
         } else {
-            imageViewFace.setImageResource(R.drawable.app_icon)
-            imageViewFace.visibility = View.VISIBLE
+            bindingMain.imageViewFace.setImageResource(R.drawable.app_icon)
+            bindingMain.imageViewFace.visibility = View.VISIBLE
         }
 
         text = "onContinuousFaceRecognized:\n"
@@ -1726,13 +1474,13 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
             }
         }
 
-        tvContinuousFace.text = text
+        bindingMain.tvContinuousFace.text = text
     }
 
     private fun showFaceRecognitionImage(mediaKey: String) {
         if (mediaKey.isEmpty()) {
-            imageViewFace.setImageResource(R.drawable.app_icon)
-            imageViewFace.visibility = View.INVISIBLE
+            bindingMain.imageViewFace.setImageResource(R.drawable.app_icon)
+            bindingMain.imageViewFace.visibility = View.INVISIBLE
             return
         }
         executorService.execute {
@@ -1740,8 +1488,8 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
                 robot.getInputStreamByMediaKey(ContentType.FACE_RECOGNITION_IMAGE, mediaKey)
                     ?: return@execute
             runOnUiThread {
-                imageViewFace.visibility = View.VISIBLE
-                imageViewFace.setImageBitmap(BitmapFactory.decodeStream(inputStream))
+                bindingMain.imageViewFace.visibility = View.VISIBLE
+                bindingMain.imageViewFace.setImageBitmap(BitmapFactory.decodeStream(inputStream))
                 try {
                     inputStream.close()
                 } catch (e: IOException) {
@@ -1759,19 +1507,19 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
         Log.d(tag.ifEmpty { "MainActivity" }, msg)
         if (!show) return
         runOnUiThread {
-            tvLog.gravity = Gravity.BOTTOM
-            tvLog.append("· $msg \n")
+            bindingFace.tvLog.gravity = Gravity.BOTTOM
+            bindingFace.tvLog.append("· $msg \n")
         }
     }
 
     private fun clearLog() {
-        tvLog.text = ""
-        tvContinuousFace.text = ""
-        imageViewFace.visibility = View.GONE
+        bindingFace.tvLog.text = ""
+        bindingMain.tvContinuousFace.text = ""
+        bindingMain.imageViewFace.visibility = View.GONE
     }
 
     private fun startNlu() {
-        robot.startDefaultNlu(etNlu.text.toString())
+        robot.startDefaultNlu(bindingMain.etNlu.text.toString())
     }
 
     override fun onSdkError(sdkException: SdkException) {
@@ -1791,9 +1539,9 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
 
     private fun goToPosition() {
         try {
-            val x = etX.text.toString().toFloat()
-            val y = etY.text.toString().toFloat()
-            val yaw = etYaw.text.toString().toFloat()
+            val x = bindingMain.etX.text.toString().toFloat()
+            val y = bindingMain.etY.text.toString().toFloat()
+            val yaw = bindingMain.etYaw.text.toString().toFloat()
             robot.goToPosition(Position(x, y, yaw, 0), backwards = false, noBypass = false)
         } catch (e: Exception) {
             e.printStackTrace()
@@ -1806,8 +1554,8 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
     }
 
     override fun onTtsVisualizerWaveFormDataChanged(waveForm: ByteArray) {
-        visualizerView.visibility = View.VISIBLE
-        visualizerView.updateVisualizer(waveForm)
+        bindingMain.visualizerView.visibility = View.VISIBLE
+        bindingMain.visualizerView.updateVisualizer(waveForm)
     }
 
     override fun onTtsVisualizerFftDataChanged(fft: ByteArray) {
@@ -2016,9 +1764,9 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
 
     private fun loadMapWithPosition(reposeRequired: Boolean) {
         try {
-            val x = etX.text.toString().toFloat()
-            val y = etY.text.toString().toFloat()
-            val yaw = etYaw.text.toString().toFloat()
+            val x = bindingMain.etX.text.toString().toFloat()
+            val y = bindingMain.etY.text.toString().toFloat()
+            val yaw = bindingMain.etYaw.text.toString().toFloat()
             loadMap(true, Position(x, y, yaw, 0))
             val position = Position(x, y, yaw, 0)
             loadMap(reposeRequired, position)
@@ -2169,6 +1917,7 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
                 )
             }
 
+            @Deprecated("Deprecated in Java")
             override fun onError(utteranceId: String?) {
                 printLog("custom tts error")
                 if (mTtsRequest == null || utteranceId == null) {
@@ -2235,7 +1984,7 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
 
     @SuppressLint("SetTextI18n")
     override fun onGreetModeStateChanged(state: Int) {
-        tvGreetMode.text = "Greet Mode -> ${OnGreetModeStateChangedListener.State.fromValue(state)}"
+        bindingMain.tvGreetMode.text = "Greet Mode -> ${OnGreetModeStateChangedListener.State.fromValue(state)}"
     }
 
     override fun onLoadFloorStatusChanged(status: Int) {
@@ -2248,5 +1997,13 @@ class MainActivity : AppCompatActivity(), NlpListener, OnRobotReadyListener,
 
     override fun onRobotDragStateChanged(isDragged: Boolean) {
         printLog("onRobotDragStateChanged $isDragged")
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun GreetingPreview() {
+    SdkTheme {
+        AppNavigation()
     }
 }
