@@ -1,4 +1,4 @@
-package com.robotemi.sdk.sample
+package com.seotepa.infobotApp
 
 import android.annotation.SuppressLint
 import android.graphics.Color
@@ -8,6 +8,7 @@ import androidx.appcompat.app.AppCompatActivity
 import com.robotemi.sdk.Robot
 import com.robotemi.sdk.TtsRequest
 import com.robotemi.sdk.listeners.OnSerialRawDataListener
+import com.seotepa.infobotApp.databinding.ActivitySerialBinding
 import com.robotemi.sdk.serial.Serial
 import com.robotemi.sdk.serial.Serial.cmd
 import com.robotemi.sdk.serial.Serial.dataFrame
@@ -15,25 +16,26 @@ import com.robotemi.sdk.serial.Serial.dataHex
 import com.robotemi.sdk.serial.Serial.getLcdBytes
 import com.robotemi.sdk.serial.Serial.getLcdColorBytes
 import com.robotemi.sdk.serial.Serial.weight
-import kotlinx.android.synthetic.main.activity_serial.*
 
 class SerialActivity : AppCompatActivity(), OnSerialRawDataListener {
 
     private var trayStatus = hashMapOf<Int, Boolean>()
 
+    private lateinit var bindingSerial: ActivitySerialBinding
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_serial)
+        bindingSerial = ActivitySerialBinding.inflate(layoutInflater)
+        setContentView(bindingSerial.root)
 
         Robot.getInstance().addOnSerialRawDataListener(this)
     }
 
     override fun onResume() {
-        ibBack.setOnClickListener { finish() }
-        btnVersion.setOnClickListener {
+        bindingSerial.ibBack.setOnClickListener { finish() }
+        bindingSerial.btnVersion.setOnClickListener {
             Robot.getInstance().sendSerialCommand(Serial.CMD_SYSTEM_GET_VERSION, byteArrayOf())
         }
-        btnTrayRGB.setOnClickListener {
+        bindingSerial.btnTrayRGB.setOnClickListener {
             Robot.getInstance().sendSerialCommand(
                 Serial.CMD_TRAY_LIGHT,
                 byteArrayOf(0x00, 0xff.toByte(), 0x00, 0x00)
@@ -48,7 +50,7 @@ class SerialActivity : AppCompatActivity(), OnSerialRawDataListener {
             ) // TRAY 3 BLUE
         }
 
-        btnStripConstant.setOnClickListener {
+        bindingSerial.btnStripConstant.setOnClickListener {
             val color = if (it.tag == true) {
                 it.tag = false
                 byteArrayOf(0xff.toByte(), 0x00, 0x00)  // Strip always on red
@@ -62,13 +64,13 @@ class SerialActivity : AppCompatActivity(), OnSerialRawDataListener {
             ) // Strip constant color
         }
 
-        btnStripOFF.setOnClickListener {
+        bindingSerial.btnStripOFF.setOnClickListener {
             Robot.getInstance().sendSerialCommand(
                 Serial.CMD_STRIP_LIGHT,
                 Serial.getStripBytes(mode = 1, primaryColor = byteArrayOf(0x00, 0x00, 0x00))
             ) // Strip always on Black, meaning turn off LED strip.
         }
-        btnStripBreathing.setOnClickListener {
+        bindingSerial.btnStripBreathing.setOnClickListener {
             val (primaryColor, secondaryColor) = if (it.tag == true) {
                 it.tag = false
                 byteArrayOf(0xff.toByte(), 0x00, 0x00) to byteArrayOf(
@@ -94,7 +96,7 @@ class SerialActivity : AppCompatActivity(), OnSerialRawDataListener {
                 )
             ) // Strip breathing
         }
-        btnStripRunning.setOnClickListener {
+        bindingSerial.btnStripRunning.setOnClickListener {
             val (primaryColor, secondaryColor) = if (it.tag == true) {
                 it.tag = false
                 byteArrayOf(0xff.toByte(), 0x00, 0x00) to byteArrayOf(
@@ -123,21 +125,21 @@ class SerialActivity : AppCompatActivity(), OnSerialRawDataListener {
             ) // Strip running
         }
 
-        btnTrayQuery.setOnClickListener {
+        bindingSerial.btnTrayQuery.setOnClickListener {
             Robot.getInstance().sendSerialCommand(
                 Serial.CMD_TRAY_SENSOR,
                 byteArrayOf(0x00)
             )
         }
 
-        btnLcdTextTime.setOnClickListener {
+        bindingSerial.btnLcdTextTime.setOnClickListener {
             Robot.getInstance().sendSerialCommand(
                 Serial.CMD_LCD_TEXT,
                 getLcdBytes(System.currentTimeMillis().toString().substring(5))
             )
         }
 
-        btnLcdTextColor.setOnClickListener {
+        bindingSerial.btnLcdTextColor.setOnClickListener {
             val color = if (it.tag == true) {
                 it.tag = false
                 byteArrayOf(0xFF.toByte(), 0x00, 0x00)
@@ -151,7 +153,7 @@ class SerialActivity : AppCompatActivity(), OnSerialRawDataListener {
             )
         }
 
-        btnLcdBackgroundColor.setOnClickListener {
+        bindingSerial.btnLcdBackgroundColor.setOnClickListener {
             val color = if (it.tag == true) {
                 it.tag = false
                 byteArrayOf(0x00, 0xFF.toByte(), 0x00)
@@ -183,7 +185,7 @@ class SerialActivity : AppCompatActivity(), OnSerialRawDataListener {
 
         // To see the hex array of raw data
         Log.d("Serial", "cmd $cmd raw data ${data.dataHex}")
-        tvResp.text = data.dataHex.toString()
+        bindingSerial.tvResp.text = data.dataHex.toString()
         when (cmd) {
             Serial.RESP_TRAY_SENSOR -> {
                 // The first place in data frame stands for tray number, starts from 0
@@ -241,7 +243,7 @@ class SerialActivity : AppCompatActivity(), OnSerialRawDataListener {
             Serial.RESP_SYSTEM_VERSION -> {
                 val decode = dataFrame.decodeToString()
                 Log.d("Serial", "decode $decode")
-                btnVersion.text = "Version:${dataFrame.decodeToString()}"
+                bindingSerial.btnVersion.text = "Version:${dataFrame.decodeToString()}"
             }
         }
     }
@@ -259,10 +261,10 @@ class SerialActivity : AppCompatActivity(), OnSerialRawDataListener {
             Color.WHITE
         }
         val view = when (trayNum) {
-            1 -> tvWeight1
-            2 -> tvWeight2
+            1 -> bindingSerial.tvWeight1
+            2 -> bindingSerial.tvWeight2
             else -> {
-                tvWeight3
+                bindingSerial.tvWeight3
             }
         }
         view.post {
